requirements.txt
django
djangorestframework
django-sslserver